# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files/folders/{folder}/files": {
      "post": {
        "description": "Upload single or multiple files to a specified folder",
        "operationId": "FileController_uploadFiles",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "path",
            "description": "Target folder name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "required": false,
            "in": "query",
            "description": "Optional custom filename for single file upload",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Files to upload"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Upload files to a folder",
        "tags": [
          "Folders & Files"
        ]
      },
      "get": {
        "operationId": "FileController_listFiles",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "path",
            "description": "Folder name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all files in a folder",
        "tags": [
          "Folders & Files"
        ]
      }
    },
    "/files/folders": {
      "get": {
        "operationId": "FileController_listFolders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Folders listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error while listing folders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "List all folders",
        "tags": [
          "Folders & Files",
          "Folder Management"
        ]
      },
      "post": {
        "operationId": "FileController_createFolder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "folderName": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Folder created successfully"
          },
          "400": {
            "description": "Invalid folder name"
          }
        },
        "summary": "Create a new folder",
        "tags": [
          "Folders & Files",
          "Folder Management"
        ]
      }
    },
    "/files/folders/{folder}": {
      "get": {
        "operationId": "FileController_getFolderDetails",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "path",
            "description": "Folder name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of files per page",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Folder details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderDetailsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Folder not found"
          }
        },
        "summary": "Get folder details and list files",
        "tags": [
          "Folders & Files"
        ]
      },
      "delete": {
        "operationId": "FileController_deleteFolder",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "path",
            "description": "Folder to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a folder and all its contents",
        "tags": [
          "Folders & Files",
          "Folder Management"
        ]
      }
    },
    "/files/folders/{folder}/files/{filename}/download": {
      "get": {
        "operationId": "FileController_downloadFile",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "path",
            "description": "Folder name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "description": "File name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Download a file from a folder",
        "tags": [
          "Folders & Files",
          "File Operations"
        ]
      }
    },
    "/files/folders/{folder}/files/{filename}/metadata": {
      "get": {
        "operationId": "FileController_getFileMetadata",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "path",
            "description": "Folder name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "description": "File name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File metadata retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileMetadataResponse"
                }
              }
            }
          },
          "404": {
            "description": "File not found"
          }
        },
        "summary": "Get metadata of a file",
        "tags": [
          "Folders & Files",
          "File Metadata"
        ]
      },
      "put": {
        "operationId": "FileController_updateFileMetadata",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "path",
            "description": "Folder name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "description": "File name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newFilename": {
                    "type": "string",
                    "description": "New filename"
                  },
                  "newFolder": {
                    "type": "string",
                    "description": "New folder"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File metadata updated successfully"
          },
          "400": {
            "description": "Invalid metadata"
          },
          "404": {
            "description": "File not found"
          }
        },
        "summary": "Update file metadata",
        "tags": [
          "Folders & Files",
          "File Metadata"
        ]
      }
    },
...
    "/files/folders/{folder}/files/download-all": {
      "get": {
        "operationId": "FileController_downloadAllFiles",
        "parameters": [
          {
            "name": "folder",
            "required": true,
            "in": "path",
            "description": "Folder name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Download all files in a folder as a ZIP archive",
        "tags": [
          "Folders & Files"
        ]
      }
    },
...
  },
  "info": {
    "title": "File Storage API",
    "description": "API for managing files and folders in the Asset Management System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "FolderResponse": {
        "type": "object",
        "properties": {
          "folders": {
            "example": [
              "folder1",
              "folder2"
            ],
            "description": "List of folder names",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "folders"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "File not found",
            "description": "Error message"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Error type"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "FileMetadataResponse": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "example": "document.pdf",
            "description": "Name of the file"
          },
          "size": {
            "type": "number",
            "example": 1024,
            "description": "Size of file in bytes"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-02-20T10:00:00.000Z",
            "description": "Creation timestamp"
          },
          "modifiedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-02-20T11:30:00.000Z",
            "description": "Last modification timestamp"
          }
        },
        "required": [
          "filename",
          "size",
          "createdAt",
          "modifiedAt"
        ]
      },
...
    }
  }
}